Make dynamic file fields work on contributions and transactions
closes #1697
